+++ reservation-service/src/main/java/ru/oleg/rsoi/service/reservation/service/SeanceServiceImpl.java
@@ -61,7 +61,9 @@ public class SeanceServiceImpl implements SeanceService {
 
     @Override
     public void deleteSeance(Integer id) {
-        seanceRepository.delete(id);
+        if (seanceRepository.exists(id)) {
+            seanceRepository.delete(id);
+        }
     }
 
     private void createSeatsForSeance(Seance seance) {
